requirements.txt
# Core dependencies for TradingAgents framework
strands>=0.1.0
langchain>=0.1.0
openai>=1.0.0
anthropic>=0.20.0
boto3>=1.34.0
botocore>=1.34.0

# Data processing and analysis
pandas>=2.0.0
numpy>=1.24.0
yfinance>=0.2.0
stockstats>=0.6.0

# Web scraping and APIs
requests>=2.31.0
beautifulsoup4>=4.12.0
finnhub-python>=2.4.0
praw>=7.7.0  # Reddit API

# Date and time handling
python-dateutil>=2.8.0

# Configuration and environment
python-dotenv>=1.0.0
pydantic>=2.0.0

# CLI and UI
typer>=0.9.0
rich>=13.0.0
questionary>=2.0.0

# Database and storage
chromadb>=0.4.0

# Async and concurrency
asyncio-mqtt>=0.13.0

# Development and testing
pytest>=7.0.0
pytest-asyncio>=0.21.0

# Optional: For enhanced functionality
tqdm>=4.65.0  # Progress bars
lxml>=4.9.0   # XML/HTML parsing